@@ -17,7 +17,7 @@ import {AutoRefreshService} from '../../services/auto-refresh.service';
 @Component({
   selector: 'wm-mappings',
   templateUrl: './mappings.component.html',
-  styleUrls: ['./mappings.component.scss']
+  styleUrls: [ './mappings.component.scss' ]
 })
 export class MappingsComponent implements OnInit, OnDestroy, WebSocketListener {
 
@@ -27,7 +27,7 @@ export class MappingsComponent implements OnInit, OnDestroy, WebSocketListener {
 
   @HostBinding('class') classes = 'wmHolyGrailBody';
 
-  @ViewChild('editor') editor;
+  @ViewChild('editor', {static: false}) editor;
 
   private ngUnsubscribe: Subject<any> = new Subject();
 
@@ -333,7 +333,7 @@ export class MappingsComponent implements OnInit, OnDestroy, WebSocketListener {
 
   editViaKeyboard($event, activeItem) {
     if (activeItem != null && (!activeItem.isProxy() || activeItem.isProxyEnabled())) {
-      if(this.editMode === State.NORMAL) {
+      if (this.editMode === State.NORMAL) {
         this.editMapping(activeItem);
       }
     }
@@ -343,7 +343,7 @@ export class MappingsComponent implements OnInit, OnDestroy, WebSocketListener {
   }
 
   abortViaKeyboard($event) {
-    if(this.editMode === State.EDIT || this.editMode === State.NEW){
+    if (this.editMode === State.EDIT || this.editMode === State.NEW) {
       this.editMode = State.NORMAL;
     }
 
@@ -353,9 +353,9 @@ export class MappingsComponent implements OnInit, OnDestroy, WebSocketListener {
 
   saveViaKeyboard($event, activeItem) {
     if (activeItem != null) {
-      if(this.editMode === State.NEW){
+      if (this.editMode === State.NEW) {
         this.saveNewMapping();
-      }  else if(this.editMode === State.EDIT){
+      } else if (this.editMode === State.EDIT) {
         this.saveEditMapping(activeItem);
       }
     }
