@@ -1,19 +1,19 @@
 import {Component, HostBinding, Input, OnChanges, OnInit, SimpleChanges} from '@angular/core';
 import {Item} from '../../model/wiremock/item';
 import {ServeEvent} from '../../model/wiremock/serve-event';
-import {UtilService} from "../../services/util.service";
-import {StubMapping} from "../../model/wiremock/stub-mapping";
-import {ResponseDefinition} from "../../model/wiremock/response-definition";
-import {WiremockService} from "../../services/wiremock.service";
-import {debounceTime} from "rxjs/operators";
-import {LoggedRequest} from "../../model/wiremock/logged-request";
-import * as qs from "querystring";
+import {UtilService} from '../../services/util.service';
+import {StubMapping} from '../../model/wiremock/stub-mapping';
+import {ResponseDefinition} from '../../model/wiremock/response-definition';
+import {WiremockService} from '../../services/wiremock.service';
+import {debounceTime} from 'rxjs/operators';
+import {LoggedRequest} from '../../model/wiremock/logged-request';
+import * as qs from 'querystring';
 
 
 @Component({
   selector: 'wm-separated',
   templateUrl: './separated.component.html',
-  styleUrls: ['./separated.component.scss']
+  styleUrls: [ './separated.component.scss' ]
 })
 export class SeparatedComponent implements OnInit, OnChanges {
 
@@ -21,7 +21,7 @@ export class SeparatedComponent implements OnInit, OnChanges {
 
   private _activeItem: Item;
 
-  color: string[] = ['bg-warning', 'bg-info', 'bg-danger', 'bg-primary', 'bg-secondary', 'bg-dark'];
+  color: string[] = [ 'bg-warning', 'bg-info', 'bg-danger', 'bg-primary', 'bg-secondary', 'bg-dark' ];
 
   bodyFileData: string;
   bodyGroupKey: string;
@@ -34,7 +34,7 @@ export class SeparatedComponent implements OnInit, OnChanges {
 
   @Input()
   set activeItem(value: Item) {
-    if(UtilService.isUndefined(this._activeItem) || this._activeItem.getCode() != value.getCode()){
+    if (UtilService.isUndefined(this._activeItem) || this._activeItem.getCode() !== value.getCode()) {
       this._activeItem = value;
     }
   }
@@ -51,17 +51,18 @@ export class SeparatedComponent implements OnInit, OnChanges {
       if (this._activeItem instanceof StubMapping) {
         responseDefinition = (this._activeItem as StubMapping).response;
         this.bodyGroupKey = 'response';
-      }else if (this._activeItem instanceof ServeEvent) {
+      } else if (this._activeItem instanceof ServeEvent) {
         responseDefinition = (this._activeItem as ServeEvent).responseDefinition;
         this.bodyGroupKey = 'responseDefinition';
-      }else{
+      } else {
         responseDefinition = null;
       }
 
       // body from file
       if (UtilService.isDefined(responseDefinition) && UtilService.isDefined(responseDefinition.bodyFileName)) {
-        this.wiremockService.getFileBody(responseDefinition.bodyFileName).pipe(debounceTime(500)).subscribe(body => this.bodyFileData = body);
-      }else{
+        this.wiremockService.getFileBody(responseDefinition.bodyFileName)
+          .pipe(debounceTime(500)).subscribe(body => this.bodyFileData = body);
+      } else {
         this.bodyFileData = null;
         this.bodyGroupKey = null;
       }
@@ -76,15 +77,15 @@ export class SeparatedComponent implements OnInit, OnChanges {
     let headers = {};
     let body;
     if (UtilService.isDefined(responseDefinition)) {
-        headers = responseDefinition.headers;
-        body = responseDefinition.body;
+      headers = responseDefinition.headers;
+      body = responseDefinition.body;
     } else if (this._activeItem instanceof LoggedRequest) {
       headers = (this._activeItem as LoggedRequest).headers;
       body = (this._activeItem as LoggedRequest).body;
     }
 
-    if(UtilService.isDefined(headers) && UtilService.isDefined(headers['Content-Type'])
-      && headers['Content-Type'] === 'application/x-www-form-urlencoded'){
+    if (UtilService.isDefined(headers) && UtilService.isDefined(headers['Content-Type'])
+      && headers['Content-Type'] === 'application/x-www-form-urlencoded') {
       // found x-www-form-urlencoded. Try to check body
       this.xWwwFormUrlEncodedParams = JSON.stringify(qs.parse(body));
     } else {
