@@ -3,7 +3,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {MappingsComponent} from './components/mappings/mappings.component';
 import {MatchedComponent} from './components/matched/matched.component';
 import {UnmatchedComponent} from './components/unmatched/unmatched.component';
-import {StateComponent} from "./components/state/state.component";
+import {StateComponent} from './components/state/state.component';
 
 const routes: Routes = [
   {
@@ -24,8 +24,8 @@ const routes: Routes = [
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  imports: [ RouterModule.forRoot(routes) ],
+  exports: [ RouterModule ]
 })
 export class AppRoutingModule {
 }
