@@ -3,7 +3,7 @@ import {Component, HostBinding, Input, OnInit} from '@angular/core';
 @Component({
   selector: 'wm-code-entry',
   templateUrl: './code-entry.component.html',
-  styleUrls: ['./code-entry.component.scss']
+  styleUrls: [ './code-entry.component.scss' ]
 })
 export class CodeEntryComponent implements OnInit {
 
