@@ -6,7 +6,7 @@ import {UtilService} from '../../services/util.service';
   // selector: 'wm-message',
   selector: 'wm-message',
   templateUrl: './message.component.html',
-  styleUrls: ['./message.component.scss']
+  styleUrls: [ './message.component.scss' ]
 })
 export class MessageComponent implements OnInit {
 
