@@ -4,7 +4,7 @@ import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 @Component({
   selector: 'wm-dialog-recording',
   templateUrl: './dialog-recording.component.html',
-  styleUrls: ['./dialog-recording.component.scss']
+  styleUrls: [ './dialog-recording.component.scss' ]
 })
 export class DialogRecordingComponent implements OnInit {
 
