@@ -1,5 +1,5 @@
-import {StubMapping} from "./wiremock/stub-mapping";
-import {dia} from "jointjs";
+import {StubMapping} from './wiremock/stub-mapping';
+import {dia} from 'jointjs';
 import Link = dia.Link;
 
 export class StateLink {
