@@ -1,5 +1,5 @@
-import { Injectable } from '@angular/core';
-import {BehaviorSubject} from "rxjs/internal/BehaviorSubject";
+import {Injectable} from '@angular/core';
+import {BehaviorSubject} from 'rxjs/internal/BehaviorSubject';
 
 @Injectable({
   providedIn: 'root'
