@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Subject} from "rxjs/internal/Subject";
+import {Subject} from 'rxjs/internal/Subject';
 
 @Injectable({
   providedIn: 'root'
